refactor: extract block-span logic from anydoc_placement into anydoc_spans (#307)
Moves _block_spans and _merge_continuations into a dedicated
anydoc_spans module to bring anydoc_placement.py under the 300-line
source file limit (310 → 263 lines).
