+++ obsidian_import/anydoc_placement.py
@@ -29,12 +29,12 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from anydoc import Block, Document, Inline
 
+from obsidian_import.anydoc_spans import block_spans
 from obsidian_import.extraction_result import MediaFile
 from obsidian_import.formatting import make_media_wikilink
 
 log = logging.getLogger(__name__)
 
-_FENCE = "```"
 _ALPHANUMERIC = re.compile(r"[^0-9a-z]+")
 # How much of a block's own text must prefix the markdown block it is matched
 # against. Long enough that neighbouring blocks cannot be confused, short
@@ -67,7 +67,7 @@ def place_media_embeds(
     Images whose position cannot be established are left out; the caller is
     responsible for surfacing them (``extract_file`` appends the leftovers).
     """
-    spans = _block_spans(markdown)
+    spans = block_spans(markdown)
     insertions = _plan_insertions(document, spans, markdown)
 
     additions: list[tuple[int, str]] = []
@@ -82,53 +82,6 @@ def place_media_embeds(
             additions.append((offset, f"{paragraph}\n\n" if at_block_start else f"\n\n{paragraph}"))
 
     return _apply_insertions(markdown, additions)
-
-
-def _block_spans(markdown: str) -> list[tuple[int, int]]:
-    """Return (start, end) character spans of the blank-line separated markdown blocks.
-
-    Fenced code blocks are kept whole: the blank lines inside them do not
-    separate blocks.
-    """
-    spans: list[tuple[int, int]] = []
-    offset = 0
-    start: int | None = None
-    end = 0
-    in_fence = False
-
-    for line in markdown.splitlines(keepends=True):
-        stripped = line.strip()
-        if stripped.startswith(_FENCE):
-            in_fence = not in_fence
-        if not stripped and not in_fence:
-            if start is not None:
-                spans.append((start, end))
-                start = None
-        else:
-            if start is None:
-                start = offset
-            end = offset + len(line.rstrip("\n"))
-        offset += len(line)
-
-    if start is not None:
-        spans.append((start, end))
-    return _merge_continuations(spans, markdown)
-
-
-def _merge_continuations(spans: list[tuple[int, int]], markdown: str) -> list[tuple[int, int]]:
-    """Join a span onto the previous one when it is an indented continuation of it.
-
-    A nested list is one block of the document model but anydoc renders it with
-    a blank line before the nested items (`1. Outer\\n\\n   1. Inner`), which
-    would otherwise split one block across two spans.
-    """
-    merged: list[tuple[int, int]] = []
-    for start, end in spans:
-        if merged and markdown[start] in " \t":
-            merged[-1] = (merged[-1][0], end)
-        else:
-            merged.append((start, end))
-    return merged
 
 
 def _plan_insertions(
